tests/python/sdk/langchain/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
distro==1.9.0
galileo>=0.0.7
galileo-core==3.11.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain>=0.1.0
langchain-core==0.3.44
langchain-openai>=0.0.2
langchain-text-splitters==0.3.6
langsmith==0.3.13
openai>=1.0.0
orjson==3.10.15
packaging==24.2
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv>=0.19.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.39
tenacity==9.0.0
tiktoken==0.9.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
wrapt==1.17.2
zstandard==0.23.0
rich>=12.0.0
